Blogilistan laajennus, tehtävät 4.15-4.23

+++ controllers/blogs.js
@@ -1,22 +1,34 @@
 const blogsRouter = require('express').Router()
+const middleware = require('../utils/middleware')
+const userExtractor = middleware.userExtractor
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  response.json(blog)
+})
+
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+  const user = await User.findById(request.user.id)
 
   const blog = new Blog({
     title: body.title,
     author: body.author || 'N/A',
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user.id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
@@ -30,9 +42,16 @@ blogsRouter.put('/:id', async (request, response) => {
   response.json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const userId = request.user.id
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === userId) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'user not authorized' })
+  }
 })
 
 blogsRouter.get('/info', async (request, response, next) => {
